VirtualShop/VirtualShop/src/Application/Products/Commands/DeleteProduct/DeleteProduct.cs
using VirtualShop.Application.Common.Models;

namespace VirtualShop.Application.Products.Commands.DeleteProduct;

public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand, Result>
{
    private readonly IProductRepository _productRepository;

    public DeleteProductCommandHandler(IProductRepository productRepository)
    {
        _productRepository = productRepository;
    }

    public async Task<Result> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
    {
        var product = await _productRepository.GetByIdAsync(request.Id);
        if (product is null)
        {
            return Result.Failure(["no such product!"]);
        }
        await _productRepository.DeleteAsync(product);

        return Result.Success();
    }
}
